packages/f8pystudio/pyproject.toml
[project]
name = "f8pystudio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "qtpy>=2.4.3",
    "pyside6>=6.10.1",
    "setuptools",
    "nodegraphqt>=0.6.43",
    "pyqtgraph>=0.14.0",
    "nats-py>=2.12.0",
    "datamodel-code-generator>=0.50.0",
    "shortuuid>=1.0.13",
    "qtawesome>=1.4.1",
    "jedi>=0.19.2",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.11.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["f8pystudio"]
include = [
    "f8pystudio/render_nodes/web_assets/**/*.html",
    "f8pystudio/render_nodes/web_assets/**/*.js",
    "f8pystudio/render_nodes/web_assets/**/*.css",
]

[tool.uv]
package = true

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
show_error_codes = true
pretty = true
follow_imports = "skip"
ignore_missing_imports = true
files = [
    "f8pystudio/bridge",
    "f8pystudio/extensions",
    "f8pystudio/session_migration.py",
    "f8pystudio/widgets/state_controls",
]

[[tool.mypy.overrides]]
module = [
    "f8pystudio.nodegraph.service_basenode",
    "f8pystudio.nodegraph.operator_basenode",
    "f8pystudio.nodegraph.node_graph",
    "f8pystudio.widgets.node_property_widgets",
    "f8pystudio.pystudio_service_bridge",
]
ignore_errors = true
